fix int argumets, refractor of ArgumentTypes

+++ jage-gpu-core/src/main/java/org/jage/gpu/agent/GpuReader.java
@@ -3,5 +3,5 @@ package org.jage.gpu.agent;
 public interface GpuReader {
     double readDouble() ;
 
-    double readInt();
+    int readInt();
 }
